+++ init.lua
@@ -15,9 +15,13 @@ local modpath = minetest.get_modpath("moreblocks")
 moreblocks.S = minetest.get_translator("moreblocks")
 
 dofile(modpath .. "/config.lua")
+dofile(modpath .. "/sounds.lua")
 dofile(modpath .. "/circular_saw.lua")
 dofile(modpath .. "/stairsplus/init.lua")
-dofile(modpath .. "/nodes.lua")
-dofile(modpath .. "/redefinitions.lua")
-dofile(modpath .. "/crafting.lua")
-dofile(modpath .. "/aliases.lua")
+
+if minetest.get_modpath("default") then
+    dofile(modpath .. "/nodes.lua")
+    dofile(modpath .. "/redefinitions.lua")
+    dofile(modpath .. "/crafting.lua")
+    dofile(modpath .. "/aliases.lua")
+end