+++ nodes.lua
@@ -7,15 +7,15 @@ Licensed under the zlib license. See LICENSE.md for more information.
 
 local S = moreblocks.S
 
-local sound_dirt = default.node_sound_dirt_defaults()
-local sound_wood = default.node_sound_wood_defaults()
-local sound_stone = default.node_sound_stone_defaults()
-local sound_glass = default.node_sound_glass_defaults()
-local sound_leaves = default.node_sound_leaves_defaults()
+local sound_dirt = moreblocks.node_sound_dirt_defaults()
+local sound_wood = moreblocks.node_sound_wood_defaults()
+local sound_stone = moreblocks.node_sound_stone_defaults()
+local sound_glass = moreblocks.node_sound_glass_defaults()
+local sound_leaves = moreblocks.node_sound_leaves_defaults()
 
 -- Don't break on 0.4.14 and earlier.
-local sound_metal = (default.node_sound_metal_defaults
-		and default.node_sound_metal_defaults() or sound_stone)
+local sound_metal = (moreblocks.node_sound_metal_defaults
+		and moreblocks.node_sound_metal_defaults() or sound_stone)
 
 local function tile_tiles(name)
 	local tex = "moreblocks_" ..name.. ".png"
@@ -175,6 +175,12 @@ local nodes = {
 		is_ground_content = false,
 		sounds = sound_stone,
 	},
+	["desert_cobble_compressed"] = {
+		description = S("Compressed Desert Cobblestone"),
+		groups = {cracky = 1},
+		is_ground_content = false,
+		sounds = sound_stone,
+	},
 	["plankstone"] = {
 		description = S("Plankstone"),
 		paramtype2 = "facedir",
@@ -188,6 +194,7 @@ local nodes = {
 		description = S("Iron Glass"),
 		drawtype = "glasslike_framed_optional",
 		tiles = {"default_glass.png^[colorize:#DEDEDE", "default_glass_detail.png^[colorize:#DEDEDE"},
+		use_texture_alpha = "clip",
 		paramtype = "light",
 		sunlight_propagates = true,
 		is_ground_content = false,
@@ -198,6 +205,7 @@ local nodes = {
 		description = S("Coal Glass"),
 		drawtype = "glasslike_framed_optional",
 		tiles = {"default_glass.png^[colorize:#828282", "default_glass_detail.png^[colorize:#828282"},
+		use_texture_alpha = "clip",
 		paramtype = "light",
 		sunlight_propagates = true,
 		is_ground_content = false,
@@ -208,6 +216,7 @@ local nodes = {
 		description = S("Clean Glass"),
 		drawtype = "glasslike_framed_optional",
 		tiles = {"moreblocks_clean_glass.png", "moreblocks_clean_glass_detail.png"},
+		use_texture_alpha = "clip",
 		paramtype = "light",
 		sunlight_propagates = true,
 		is_ground_content = false,
@@ -307,6 +316,7 @@ local nodes = {
 		description = S("Trap Glass"),
 		drawtype = "glasslike_framed_optional",
 		tiles = {"default_glass.png^moreblocks_trap_box_glass.png", "default_glass_detail.png"},
+		use_texture_alpha = "clip",
 		paramtype = "light",
 		sunlight_propagates = true,
 		is_ground_content = false,
@@ -319,6 +329,7 @@ local nodes = {
 		description = S("Trap Obsidian Glass"),
 		drawtype = "glasslike_framed_optional",
 		tiles = {"default_obsidian_glass.png^moreblocks_trap_box_glass.png", "default_obsidian_glass_detail.png"},
+		use_texture_alpha = "clip",
 		paramtype = "light",
 		sunlight_propagates = true,
 		is_ground_content = false,
@@ -336,6 +347,19 @@ local nodes = {
 		paramtype = "light",
 		is_ground_content = false,
 		sounds = sound_stone,
+		no_stairs = true,
+	},
+	["trap_clean_glass"] = {
+		description = S("Trap Clean Glass"),
+		drawtype = "glasslike_framed_optional",
+		tiles = {"moreblocks_clean_glass.png^moreblocks_trap_box_glass.png", "moreblocks_clean_glass_detail.png"},
+		use_texture_alpha = "clip",
+		paramtype = "light",
+		sunlight_propagates = true,
+		is_ground_content = false,
+		walkable = false,
+		groups = {cracky = 3, oddly_breakable_by_hand = 3},
+		sounds = sound_glass,
 		no_stairs = true,
 	},
 	["trap_sandstone"] = {
@@ -388,6 +412,19 @@ local nodes = {
 		description = S("Glow Glass"),
 		drawtype = "glasslike_framed_optional",
 		tiles = {"default_glass.png^[colorize:#E9CD61", "default_glass_detail.png^[colorize:#E9CD61"},
+		use_texture_alpha = "clip",
+		paramtype = "light",
+		sunlight_propagates = true,
+		is_ground_content = false,
+		light_source = 11,
+		groups = {cracky = 3, oddly_breakable_by_hand = 3},
+		sounds = sound_glass,
+	},
+	["clean_glow_glass"] = {
+		description = S("Clean Glow Glass"),
+		drawtype = "glasslike_framed_optional",
+		tiles = {"moreblocks_clean_glass.png^[colorize:#E9CD61", "moreblocks_clean_glass_detail.png^[colorize:#E9CD61"},
+		use_texture_alpha = "clip",
 		paramtype = "light",
 		sunlight_propagates = true,
 		is_ground_content = false,
@@ -399,6 +436,21 @@ local nodes = {
 		description = S("Trap Glow Glass"),
 		drawtype = "glasslike_framed_optional",
 		tiles = {"default_glass.png^[colorize:#E9CD61^moreblocks_trap_box_glass.png", "default_glass_detail.png^[colorize:#E9CD61"},
+		use_texture_alpha = "clip",
+		paramtype = "light",
+		sunlight_propagates = true,
+		is_ground_content = false,
+		light_source = 11,
+		walkable = false,
+		groups = {cracky = 3, oddly_breakable_by_hand = 3},
+		sounds = sound_glass,
+		no_stairs = true,
+	},
+	["trap_clean_glow_glass"] = {
+		description = S("Trap Clean Glow Glass"),
+		drawtype = "glasslike_framed_optional",
+		tiles = {"moreblocks_clean_glass.png^[colorize:#E9CD61^moreblocks_trap_box_glass.png", "moreblocks_clean_glass_detail.png^[colorize:#E9CD61"},
+		use_texture_alpha = "clip",
 		paramtype = "light",
 		sunlight_propagates = true,
 		is_ground_content = false,
@@ -412,6 +464,19 @@ local nodes = {
 		description = S("Super Glow Glass"),
 		drawtype = "glasslike_framed_optional",
 		tiles = {"default_glass.png^[colorize:#FFFF78", "default_glass_detail.png^[colorize:#FFFF78"},
+		use_texture_alpha = "clip",
+		paramtype = "light",
+		sunlight_propagates = true,
+		is_ground_content = false,
+		light_source = default.LIGHT_MAX,
+		groups = {cracky = 3, oddly_breakable_by_hand = 3},
+		sounds = sound_glass,
+	},
+	["clean_super_glow_glass"] = {
+		description = S("Clean Super Glow Glass"),
+		drawtype = "glasslike_framed_optional",
+		tiles = {"moreblocks_clean_glass.png^[colorize:#FFFF78", "moreblocks_clean_glass_detail.png^[colorize:#FFFF78"},
+		use_texture_alpha = "clip",
 		paramtype = "light",
 		sunlight_propagates = true,
 		is_ground_content = false,
@@ -423,6 +488,21 @@ local nodes = {
 		description = S("Trap Super Glow Glass"),
 		drawtype = "glasslike_framed_optional",
 		tiles = {"default_glass.png^[colorize:#FFFF78^moreblocks_trap_box_glass.png", "default_glass_detail.png^[colorize:#FFFF78"},
+		use_texture_alpha = "clip",
+		paramtype = "light",
+		sunlight_propagates = true,
+		is_ground_content = false,
+		light_source = default.LIGHT_MAX,
+		walkable = false,
+		groups = {cracky = 3, oddly_breakable_by_hand = 3},
+		sounds = sound_glass,
+		no_stairs = true,
+	},
+	["trap_clean_super_glow_glass"] = {
+		description = S("Trap Clean Super Glow Glass"),
+		drawtype = "glasslike_framed_optional",
+		tiles = {"moreblocks_clean_glass.png^[colorize:#FFFF78^moreblocks_trap_box_glass.png", "moreblocks_clean_glass_detail.png^[colorize:#FFFF78"},
+		use_texture_alpha = "clip",
 		paramtype = "light",
 		sunlight_propagates = true,
 		is_ground_content = false,